finish project

+++ src/Form/Trick/TrickHandler.php
@@ -4,12 +4,12 @@ namespace App\Form\Trick;
 
 use App\Entity\Trick;
 use App\Entity\Image;
-use App\Entity\Video;
 use App\Services\PictureService;
+use Doctrine\DBAL\Exception\DriverException;
+use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\Form\FormInterface;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
@@ -39,6 +39,14 @@ final class TrickHandler
         return $this->formFactory->create(TrickFormType::class, $data, $options);
     }
 
+    /**
+     * @param FormInterface $form
+     * @param Request $request
+     * @param Trick $trick
+     * @param string $upload_directory
+     * @param bool $isEdit
+     * @return bool
+     */
     public function handle(FormInterface $form, Request $request, Trick $trick, string $upload_directory, bool $isEdit = false): bool
     {
         $form->handleRequest($request);
@@ -46,20 +54,37 @@ final class TrickHandler
         if ($form->isSubmitted() && $form->isValid()) {
             $file = $form->get('imageFile')->getData();
             $files = $form->get('images')->getData();
-            $newVideos = [
-                $form->get('newVideo')->getData(),
-                $form->get('newVideo2')->getData(),
-                $form->get('newVideo3')->getData()
-            ];
+            $secondaryPictures = $form->get('secondaryImages')->getData();
 
             $name = $form->get('name')->getData();
             $slug = $this->setSlugForName($name);
             $trick->setSlug($slug);
 
-            if (count($files) >= 4) {
-                $session = $this->requestStack->getSession();
-                $session->getFlashBag()->add('error', 'Vous ne pouvez mettre que 3 images');
-                return false;
+            foreach ($secondaryPictures as $secondary) {
+                if ($secondary->getFile() !== null) {
+                    foreach ($trick->getSecondaryImages() as $img) {
+                        if ($secondary->getName() === $img->getName()) {
+                            $newFileName = $this->pictureService->add($secondary->getFile(), '', 300, 200);
+                            try {
+                                $this->pictureService->delete($img->getName(), 'tricksImg', 300, 200);
+
+                                $this->entityManager->remove($img);
+
+                                $this->entityManager->flush();
+
+                                $image = new Image();
+                                $image->setTricks($trick)
+                                    ->setName($newFileName);
+
+                                $trick->addSecondaryImage($image);
+                                $this->entityManager->persist($image);
+                                $this->entityManager->flush();
+                            } catch (FileException $e) {
+                                error_log($e->getMessage());
+                            }
+                        }
+                    }
+                }
             }
 
             foreach ($files as $imageForm) {
@@ -80,9 +105,10 @@ final class TrickHandler
             if ($file !== null) {
                 $existingImage = $trick->getImages();
 
-                if ($existingImage !== null) {
+                if (!empty($existingImage)) {
                     $existingImagePath = $upload_directory . '/' . $existingImage;
-                    if (file_exists($existingImagePath)) {
+
+                    if (file_exists($existingImagePath) && is_file($existingImagePath)) {
                         unlink($existingImagePath);
                     }
                 }
@@ -91,22 +117,9 @@ final class TrickHandler
 
                 try {
                     $file->move($upload_directory, $newFileName);
+                    $trick->setImages($newFileName);
                 } catch (FileException $e) {
                     error_log($e->getMessage());
-                }
-
-                $trick->setImages($newFileName);
-            }
-
-            // Ajouter les vidéos
-            foreach ($newVideos as $videoUrl) {
-                if ($videoUrl) { // Assurez-vous que l'URL de la vidéo n'est pas vide
-                    $video = new Video();
-                    $video->setTricks($trick)
-                        ->setUrl($videoUrl);
-
-                    $trick->addVideo($video);
-                    $this->entityManager->persist($video);
                 }
             }
 
@@ -119,8 +132,23 @@ final class TrickHandler
                 $trick->setCreatedAt(new \DateTimeImmutable());
             }
 
-            $this->entityManager->persist($trick);
-            $this->entityManager->flush();
+            try {
+                $this->entityManager->persist($trick);
+                $this->entityManager->flush();
+            } catch (UniqueConstraintViolationException $e) {
+                $this->requestStack->getSession()->getFlashBag()->add('error', 'Un trick avec ce nom existe déjà.');
+                error_log('UniqueConstraintViolationException: ' . $e->getMessage());
+                return false;
+            } catch (DriverException $e) {
+                // Catch database related exceptions
+                error_log('DriverException: ' . $e->getMessage());
+                $this->requestStack->getSession()->getFlashBag()->add('error', 'Erreur de base de données.');
+                return false;
+            } catch (\Exception $e) {
+                error_log('Exception: ' . $e->getMessage());
+                $this->requestStack->getSession()->getFlashBag()->add('error', 'Erreur inconnue lors de la sauvegarde du trick.');
+                return false;
+            }
 
             return true;
         }
